Remove rule update/install callouts
